chore(api-service,dashboard,dal): adjust the partner integrations rest apis and consolidate logic
