+++ apps/api/src/app/partner-integrations/dtos/create-vercel-integration-response.dto.ts
@@ -1,6 +1,6 @@
 import { IsDefined } from 'class-validator';
 
-export class SetupVercelConfigurationResponseDto {
+export class CreateVercelIntegrationResponseDto {
   @IsDefined()
   success: boolean;
 }
