--- apps/api/src/app/partner-integrations/dtos/get-vercel-projects-request.dto.ts
@@ -1,7 +1,0 @@
-import { IsDefined, IsString } from 'class-validator';
-
-export class SetVercelConfigurationRequestDto {
-  @IsDefined()
-  @IsString()
-  configurationId: string;
-}
