+++ apps/api/src/app/partner-integrations/partner-integrations.controller.ts
@@ -12,21 +12,19 @@ import {
   Headers,
 } from '@nestjs/common';
 import { UserSessionData } from '@novu/shared';
-import { ModuleRef } from '@nestjs/core';
+
 import { UserSession } from '../shared/framework/user.decorator';
-import { CompleteAndUpdateVercelIntegrationRequestDto } from './dtos/complete-and-update-vercel-integration-request.dto';
-import { SetVercelConfigurationRequestDto } from './dtos/setup-vercel-integration-request.dto';
-import { SetupVercelConfigurationResponseDto } from './dtos/setup-vercel-integration-response.dto';
-import { CompleteVercelIntegrationCommand } from './usecases/complete-vercel-integration/complete-vercel-integration.command';
-import { CompleteVercelIntegration } from './usecases/complete-vercel-integration/complete-vercel-integration.usecase';
-import { GetVercelConfigurationCommand } from './usecases/get-vercel-configuration/get-vercel-configuration.command';
-import { GetVercelConfiguration } from './usecases/get-vercel-configuration/get-vercel-configuration.usecase';
-import { GetVercelProjectsCommand } from './usecases/get-vercel-projects/get-vercel-projects.command';
-import { GetVercelProjects } from './usecases/get-vercel-projects/get-vercel-projects.usecase';
-import { SetVercelConfigurationCommand } from './usecases/set-vercel-configuration/set-vercel-configuration.command';
-import { SetVercelConfiguration } from './usecases/set-vercel-configuration/set-vercel-configuration.usecase';
-import { UpdateVercelConfigurationCommand } from './usecases/update-vercel-configuration/update-vercel-configuration.command';
-import { UpdateVercelConfiguration } from './usecases/update-vercel-configuration/update-vercel-configuration.usecase';
+import { UpdateVercelIntegrationRequestDto } from './dtos/update-vercel-integration-request.dto';
+import { CreateVercelIntegrationRequestDto } from './dtos/create-vercel-integration-request.dto';
+import { CreateVercelIntegrationResponseDto } from './dtos/create-vercel-integration-response.dto';
+import { GetVercelIntegrationCommand } from './usecases/get-vercel-integration/get-vercel-integration.command';
+import { GetVercelIntegration } from './usecases/get-vercel-integration/get-vercel-integration.usecase';
+import { GetVercelIntegrationProjectsCommand } from './usecases/get-vercel-projects/get-vercel-integration-projects.command';
+import { GetVercelIntegrationProjects } from './usecases/get-vercel-projects/get-vercel-integration-projects.usecase';
+import { CreateVercelIntegrationCommand } from './usecases/create-vercel-integration/create-vercel-integration.command';
+import { CreateVercelIntegration } from './usecases/create-vercel-integration/create-vercel-integration.usecase';
+import { UpdateVercelIntegrationCommand } from './usecases/update-vercel-integration/update-vercel-integration.command';
+import { UpdateVercelIntegration } from './usecases/update-vercel-integration/update-vercel-integration.usecase';
 import { UserAuthentication } from '../shared/framework/swagger/api.key.security';
 import { ProcessVercelWebhook } from './usecases/process-vercel-webhook/process-vercel-webhook.usecase';
 import { ProcessVercelWebhookCommand } from './usecases/process-vercel-webhook/process-vercel-webhook.command';
@@ -37,28 +35,71 @@ import { ProcessVercelWebhookCommand } from './usecases/process-vercel-webhook/p
 @ApiExcludeController()
 export class PartnerIntegrationsController {
   constructor(
-    private setVercelConfigurationUsecase: SetVercelConfiguration,
-    private getVercelProjectsUsecase: GetVercelProjects,
-    private completeVercelIntegrationUsecase: CompleteVercelIntegration,
-    private getVercelConfigurationUsecase: GetVercelConfiguration,
-    private updateVercelConfigurationUsecase: UpdateVercelConfiguration,
-    private processVercelWebhookUsecase: ProcessVercelWebhook,
-    protected moduleRef: ModuleRef
+    private createVercelIntegrationUsecase: CreateVercelIntegration,
+    private getVercelIntegrationProjectsUsecase: GetVercelIntegrationProjects,
+    private getVercelIntegrationUsecase: GetVercelIntegration,
+    private updateVercelIntegrationUsecase: UpdateVercelIntegration,
+    private processVercelWebhookUsecase: ProcessVercelWebhook
   ) {}
 
   @Post('/vercel')
   @UserAuthentication()
-  async setupVercelIntegration(
+  async createVercelIntegration(
     @UserSession() user: UserSessionData,
-    @Body() body: SetVercelConfigurationRequestDto
-  ): Promise<SetupVercelConfigurationResponseDto> {
-    return await this.setVercelConfigurationUsecase.execute(
-      SetVercelConfigurationCommand.create({
+    @Body() body: CreateVercelIntegrationRequestDto
+  ): Promise<CreateVercelIntegrationResponseDto> {
+    return await this.createVercelIntegrationUsecase.execute(
+      CreateVercelIntegrationCommand.create({
         vercelIntegrationCode: body.vercelIntegrationCode,
         environmentId: user.environmentId,
         organizationId: user.organizationId,
         userId: user._id,
         configurationId: body.configurationId,
+      })
+    );
+  }
+
+  @Put('/vercel')
+  @UserAuthentication()
+  async updateVercelIntegration(@UserSession() user: UserSessionData, @Body() body: UpdateVercelIntegrationRequestDto) {
+    return await this.updateVercelIntegrationUsecase.execute(
+      UpdateVercelIntegrationCommand.create({
+        data: body.data,
+        environmentId: user.environmentId,
+        organizationId: user.organizationId,
+        userId: user._id,
+        configurationId: body.configurationId,
+      })
+    );
+  }
+
+  @Get('/vercel/:configurationId')
+  @UserAuthentication()
+  async getVercelIntegration(@UserSession() user: UserSessionData, @Param('configurationId') configurationId: string) {
+    return await this.getVercelIntegrationUsecase.execute(
+      GetVercelIntegrationCommand.create({
+        userId: user._id,
+        configurationId,
+        environmentId: user.environmentId,
+        organizationId: user.organizationId,
+      })
+    );
+  }
+
+  @Get('/vercel/:configurationId/projects')
+  @UserAuthentication()
+  async getVercelProjects(
+    @UserSession() user: UserSessionData,
+    @Param('configurationId') configurationId: string,
+    @Query('nextPage') nextPage?: string
+  ) {
+    return await this.getVercelIntegrationProjectsUsecase.execute(
+      GetVercelIntegrationProjectsCommand.create({
+        configurationId,
+        environmentId: user.environmentId,
+        organizationId: user.organizationId,
+        userId: user._id,
+        ...(nextPage && { nextPage }),
       })
     );
   }
@@ -69,74 +110,6 @@ export class PartnerIntegrationsController {
       ProcessVercelWebhookCommand.create({
         body,
         signatureHeader,
-      })
-    );
-  }
-
-  @Get('/vercel/projects/:configurationId')
-  @UserAuthentication()
-  async getVercelProjects(
-    @UserSession() user: UserSessionData,
-    @Param('configurationId') configurationId: string,
-    @Query('nextPage') nextPage?: string
-  ) {
-    return await this.getVercelProjectsUsecase.execute(
-      GetVercelProjectsCommand.create({
-        configurationId,
-        environmentId: user.environmentId,
-        organizationId: user.organizationId,
-        userId: user._id,
-        ...(nextPage && { nextPage }),
-      })
-    );
-  }
-
-  @Post('/vercel/configuration/complete')
-  @UserAuthentication()
-  async completeVercelIntegration(
-    @UserSession() user: UserSessionData,
-    @Body() body: CompleteAndUpdateVercelIntegrationRequestDto
-  ) {
-    return await this.completeVercelIntegrationUsecase.execute(
-      CompleteVercelIntegrationCommand.create({
-        data: body.data,
-        environmentId: user.environmentId,
-        organizationId: user.organizationId,
-        userId: user._id,
-        configurationId: body.configurationId,
-      })
-    );
-  }
-
-  @Get('vercel/configuration/:configurationId')
-  @UserAuthentication()
-  async getVercelConfigurationDetails(
-    @UserSession() user: UserSessionData,
-    @Param('configurationId') configurationId: string
-  ) {
-    return await this.getVercelConfigurationUsecase.execute(
-      GetVercelConfigurationCommand.create({
-        userId: user._id,
-        configurationId,
-        environmentId: user.environmentId,
-        organizationId: user.organizationId,
-      })
-    );
-  }
-
-  @Put('/vercel/configuration/update')
-  @UserAuthentication()
-  async updateVercelConfiguration(
-    @UserSession() user: UserSessionData,
-    @Body() body: CompleteAndUpdateVercelIntegrationRequestDto
-  ) {
-    return await this.updateVercelConfigurationUsecase.execute(
-      UpdateVercelConfigurationCommand.create({
-        data: body.data,
-        environmentId: user.environmentId,
-        organizationId: user.organizationId,
-        userId: user._id,
-        configurationId: body.configurationId,
       })
     );
   }
