--- apps/api/src/app/partner-integrations/usecases/complete-vercel-integration/complete-vercel-integration.command.ts
@@ -1,6 +1,0 @@
-import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
-
-export class CompleteVercelIntegrationCommand extends EnvironmentWithUserCommand {
-  data: Record<string, string[]>;
-  configurationId: string;
-}
