+++ apps/api/src/app/partner-integrations/usecases/create-vercel-integration/create-vercel-integration.command.ts
@@ -1,7 +1,7 @@
 import { IsDefined } from 'class-validator';
 import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
 
-export class SetVercelConfigurationCommand extends EnvironmentWithUserCommand {
+export class CreateVercelIntegrationCommand extends EnvironmentWithUserCommand {
   @IsDefined()
   vercelIntegrationCode: string;
   @IsDefined()
