+++ apps/api/src/app/partner-integrations/usecases/index.ts
@@ -1,15 +1,13 @@
-import { UpdateVercelConfiguration } from './update-vercel-configuration/update-vercel-configuration.usecase';
-import { GetVercelConfiguration } from './get-vercel-configuration/get-vercel-configuration.usecase';
-import { CompleteVercelIntegration } from './complete-vercel-integration/complete-vercel-integration.usecase';
-import { GetVercelProjects } from './get-vercel-projects/get-vercel-projects.usecase';
-import { SetVercelConfiguration } from './set-vercel-configuration/set-vercel-configuration.usecase';
+import { UpdateVercelIntegration } from './update-vercel-integration/update-vercel-integration.usecase';
+import { GetVercelIntegration } from './get-vercel-integration/get-vercel-integration.usecase';
+import { GetVercelIntegrationProjects } from './get-vercel-projects/get-vercel-integration-projects.usecase';
+import { CreateVercelIntegration } from './create-vercel-integration/create-vercel-integration.usecase';
 import { ProcessVercelWebhook } from './process-vercel-webhook/process-vercel-webhook.usecase';
 
 export const USE_CASES = [
-  SetVercelConfiguration,
-  GetVercelProjects,
-  CompleteVercelIntegration,
-  GetVercelConfiguration,
-  UpdateVercelConfiguration,
+  CreateVercelIntegration,
+  GetVercelIntegrationProjects,
+  GetVercelIntegration,
+  UpdateVercelIntegration,
   ProcessVercelWebhook,
 ];
