+++ apps/api/src/app/partner-integrations/usecases/update-vercel-integration/update-vercel-integration.command.ts
@@ -1,5 +1,6 @@
 import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
 
-export class GetVercelConfigurationCommand extends EnvironmentWithUserCommand {
+export class UpdateVercelIntegrationCommand extends EnvironmentWithUserCommand {
+  data: Record<string, string[]>;
   configurationId: string;
 }
